signals not works

+++ parsing/token/ft_free_token.c
@@ -6,7 +6,7 @@
 /*   By: dcorenti <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/13 23:44:25 by dcorenti          #+#    #+#             */
-/*   Updated: 2022/11/09 22:56:30 by dcorenti         ###   ########.fr       */
+/*   Updated: 2022/11/16 18:27:57 by dcorenti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 
 */
 
-void	ft_free_token(t_token *token)
+void	ft_free_token(t_token *token, int heredoc_unlink)
 {
 	t_token	*tmp;
 
@@ -33,12 +33,12 @@ void	ft_free_token(t_token *token)
 			free(token->value);
 		if (token->red_file_name)
 		{
-			if (token->type == HEREDOC)
+			if (token->type == HEREDOC && heredoc_unlink == 1)
 				unlink(token->red_file_name);
 			free(token->red_file_name);
 		}
 		free(token);
 	}
 	if (tmp != NULL)
-		ft_free_token(tmp);
+		ft_free_token(tmp, heredoc_unlink);
 }
